wrstcse: Corrected the range random tolerances are in
wrstcse is a naive go at an interval arithmetic: An actual
interval arithmetic is bound to fail if applied to complex
equations. Wrstcse instead allows to use an undeclared
array named tol of which each element represents a tolerance
that can range between -1 and 1 and introduces
combinations of values of tol into an equation in order
to guess which range the result would be in.

If wrstcse was instructed to try random values the range of
the elements of tol didn't go from -1...1. This commit
corrects that.