PDF manual: Make hyperlinks that contain a "_" work
In the PDF version of the manual Hyperlinks to commands or
variables whose name contains an underscore ("_") never worked.

Jinsong Zhao now has found a way to patch texinfo that makes them
work. Tested on ubuntu with maxima's texinfo.tex and also the
rtoy_* branches that use new methods of generating the pdf
documentation.

Resolves #3549