Fix autoload after killing a function that caused a previous autoload
*AUTOLOADED-FILES* is a list that keeps track of which files have
been autoloaded and which function caused the autoload.  When a
function/file pair is on the list, the file is not autoloaded again
when attempting to call the function.

This list was not updated during a kill, so using the same function
again after killing it would not cause another autoload because the
function/file pair was still on the list (although using a different
function to autoload the same file would work).

Now we remove the function/file pair from the list during a kill of
the function.

This was reported by Martin Stanev on the mailing list on 2019-05-06.
The subject was "problem with elem() after kill(all)".

No problems with the test suite or share test suite.