Replace (PROGN 'COMPILE ...) with (PROGN ...)
In Maclisp and Franz Lisp the quoted COMPILE in (PROGN 'COMPILE ...)
was required at top level to tell the compiler to actually compile
the subforms as top level forms.  Treating the subforms as top level
forms in this case is standard Common Lisp, so let's get rid of the
quoted symbol that doesn't do anything useful anymore.

There were many more of these (PROGN 'COMPILE ...) forms in the past,
but they have since been changed or removed.