A workaround that makes rpm builds from inofficial tarballs work again
Maxima on creating a tarball gets its version number from
git --describe - and names the tarball and the contained files
accordingly. It also uses this name for the temp dir for rpmbuild.
But the tarball that is build for the rpm package doesn't contain
the .git directory and therefore takes the version number from
the AC_INIT() command that needs to be run before git --describe
can be issued => rpmbuild didn't know where to find its files.
Now it finds its files using wildcards - that should not be able
to pick up old builds as every build is done in its own build
directory.