Make :lisp-quiet tell :lisp which label was suppressed
This resolves #3007 except of in one case (that IMHO nothing cannot
be done about): If *prompt-prefix* or *prompt-suffix* has changed
from inside the :lisp-quiet() command that change is only effective
in the command that follows the current :lisp command.

For the sake of completeness: The case that *mread-prompt* was set
to "" by the :lisp command was neglected as if the user wanted to
do that a :lisp-quiet command the accurate means.