We forgot to ship some language-specific files in the tarball
As the

   make distcheck

make target doesn't work yet this kind of bug is to be expected
and in fact happens with nearly every release.

As this time the files were language-specific and are only
required if languages were enabled using configure arguments a
"make distcheck" wouldn't have helped, though, this time. And
as many devs build their packages from git (which doesn't have
this omission) it looks like this specific bug is more than
one maxima version old.

This commit should resolve issue #3558.