Documentation: "structures" definitively belongs to infolists.
The variable "structures" is documented in Structures.texi. Which
might be the Right Place to document it. But this variable
definitively is an infolist => Added a mention of this variable in
the "infolists" section of Command.texi, too.

If one wants to replace the "structures" documentation in
Structures.texi by a link to infolists (which introduces another
far-reaching cross-ref, but will reduce the size of the manual by
about 5 lines) I'm neither in favour nor against this.