plot: escape two more letters in function names.
Gnuplot treats { and } as the beginning and end marker for
subscripted or superscripted regions. We need to escape them,
therefore, if we want them to be part of a function name.

It may be uncommon to do so, but it is possible to define
a function with such a name in maxima:

    f\{a\}(x):=x;