Plot Docu: Made one of the examples work under all circumstances
Some front-ends by default talk XML. Since one of the plot examples
used with_stdout for creating a matrix and depending on the front-end
this might lead to an XML representation of a matrix the resulting
example might or might not work.

Setting display2d:false as the 1st line of the example makes it work
regardless of the front-end being used.