Be more careful about non-numerical input in find_root and bf_find_root
find_root and bf_find_root want to return a noun form when they
receive non-numerical input.

Now bind ERRSET and *MDEBUG* (i.e., debugmode) to NIL in the local
function CONVERT in FIND-ROOT-SUBR which attempts to convert an
expression to a bigfloat.  This is to ensure that the ERRSET macro
yields NIL when any error occurs and that we don't get dropped into
the Maxima debugger if a Maxima error occurs.

If debugmode was true (regardless of ERRSET) then things like
find_root(x,x,a,b) dropped us into the Maxima debugger without
even giving us the error message.  If ERRSET was non-NIL and
debugmode was false then a lisp error occurred instead.