Evaluate the body of a WITH-NEW-CONTEXT form as an implicit PROGN
The body has been evaluated as an implicit PROG1, which is strange
and unintuitive.  Providing an implicit PROGN is the obvious thing
to do here.  Also using PROGN instead of PROG1 is strictly more
general when multiple values are involved, but that does not affect
any of the current uses of this macro.

In Maxima 5.0 WITH-NEW-CONTEXT was implemented essentially as a
PROG1 to handle the cleanup.  There were only 3 uses of this macro
in src and each of their bodies only consisted of a single form (no
PROGN).

In commit 4b8a9864 (from 2013) this macro was rewritten so that
(among other things) it used UNWIND-PROTECT, but a PROG1 still
remained.

There are currently 8 uses of WITH-NEW-CONTEXT in src (and 0 in
share).  Every WITH-NEW-CONTEXT body has consisted of only a single
form, where 3 of these were PROGNs.  Now we can drop these explicit
PROGNs in the bodies.

No functional changes.

No problems with the test suite or share test suite.