simplimln: handle simple cases without recursion
Fixes #3587 limit(log(3-sqrt(x)),x,9,minus);
Fixes #3589 limit((sqrt(x)-2)*log(1-sqrt(x)/2),x,4,minus);

Consider the limit:
 limit(log(3-sqrt(x)),x,9,minus);
The arg of log is not real everywhere, but it is real near the limit
point.  If the arg of log is zeroa or zerob, which are both real,
we don't need to worry about the branch cut along the negative real
axis.  The recursive call on the real part and imaginary part of the
entire log expression is somewhat risky - it can cause infinite loops
because we are calling limit on an expression larger than the one we
started with.

Added link to a version of Paul Wang's thesis that has no blank pages.