Replace an incorrect comment in TOPLEVEL-OPTIMIZE
In commit 068c6102 in 2009 a comment was added to TOPLEVEL-OPTIMIZE.
It states that the code appears to be broken and that the code always
returns the symbol itself.  I think the author of the comment was
using a constantp property (mentioned in the comment) instead of the
constant property (used in the code), or perhaps they were confusing
this with the constant declaration (which they were working on in
the aforementioned commit).

I've tested the code and it seems to work fine.  If a symbol has the
constant property, then TOPLEVEL-OPTIMIZE uses that assigned constant
value in place of the symbol.  Therefore I'm updating the comment to
describe the behavior.

No functional changes.