Rewrite TRANSLATE-AND-EVAL-MACSYMA-EXPRESSION for clarity
The comment stated that the code was written the way it was so that
TRANSLATOR-EVAL and TRANSLATE-MACEXPR-TOPLEVEL could still be traced
when PURE was set to NIL.  PURE was a (PDP-10 only) Maclisp variable
which was related to read-only memory and UUO links.  See commit
47dcc8a0 for a little more on UUO links and how they're related to
tracing.

Also fix an incorrect comment in this function.

No functional changes.