Remove the special variable *BAKTRCL* and the function bt
BAKTRCL (now *BAKTRCL*) was originally a (lisp) list that kept some
backtrace information in Macsyma.  Its default value was NIL and it
was bound in MEVAL* and MEVAL1.  When DEBUG (now *MDEBUG*) was set
to all, MEVAL1 would push certain forms it was evaluating onto
BAKTRCL.  In ERRBREAK, which was an ERRSET handler, a variable named
backtrace was set to a (Maxima) list containing the elements of
BAKTRCL.

This remained the same through Maxima 5.2 (circa 1995).  In Maxima
5.4beta1 (circa 1998) src/mdebug.lisp appeared.  A function named bt
was present which could be used to print some backtrace information
held in BAKTRCL, although it seemed to be a work in progress.  Either
way, the code in MEVAL1 to push onto BAKTRCL was also commented out
and so BAKTRCL was always NIL anyway.

The commented-out code in MEVAL1 was removed in commit 66a62cf7 in
2007 and BAKTRCL was renamed *BAKTRCL* in commit f4931e42 in 2009.
Since then, *BAKTRCL* still gets bound in MEVAL* and MEVAL1 but it
is always NIL.

The function bt is undocumented and unused in src and share.  It
always just returns NIL since *BAKTRCL* is always NIL.

No problems with the test suite or share test suite.