Fix bug #3265: gcfactor(0) -> division by zero
If A and B are both 0 then their gcd as computed by GCD is 0.  We
need to check for this before dividing A and B by their gcd.

gcfactor(0) was handled correctly in Macsyma, but it looks like this
behavior changed at some point before Maxima 5.0.

No problems with the test suite or share test suite.