Remove a superfluous test in ALIAS
After commit 270ba62c in 2009, which fixed bug #464, there was a
superfluous test before signalling an error.  The (NOT (EQ X Y))
test was always true at that point because (EQ X Y) is the test
form for the previous COND clause.

I noticed this, but Stavros Macrakis actually mentioned this in
his initial report of bug #464 in 2003.

No functional changes.