Remove the special variable *PRIN1*
In Maclisp and Lisp Machine Lisp there was a variable named PRIN1
which, if non-NIL, specified a function for the toplevel REPL to call
instead of PRIN1.  This variable did not affect the output from other
functions that would call PRIN1 (like PRINT).

When non-NIL, the value of PRIN1 was explicitly called by Macsyma for
just one specific thing: printing trace output when the lisp_print
trace option was set.

Some time before Maxima 5.0 a PRIN1 variable was defined in commac.
Earmuffs were put on in commit ad12e542 in 2004.

Since the value of *PRIN1* has not been used by the underlying lisp
and its value has still only been used in this one very specific
place, I'm just removing it altogether.

No problems with the test suite or share test suite.