Bug fix: finding the free lisp variables in a LET form was broken
The FREE-LISP-VARS property for LET was broken.  When looking for
free variables in the initial value forms of the variables bound
by the LET, we had the following:

* if an initial value form was a non-symbolic atom, then a lisp
  error would occur in FREE-LISP-VARS
* if an initial value form was a symbol (a variable free in the
  LET form), then FREE-LISP-VARS would silently fail to find the
  free variable
* if an initial value form was a cons, then FREE-LISP-VARS would
  print an error message and fail to find any free variables in
  the initial value form

The issue was that CDR was used instead of CADR when grabbing the
initial value forms for the variables bound by the LET.

This was introduced in commit 95ded719 this past June 2019.