translator: improve the error message when a throw is not in a catch
I think it's awkward and unclear to just print the object being
thrown immediately before the error message.  Based solely on the
output, it's not obvious that the thing printed is the thing being
thrown:

(%i1) foo () := throw ('bar)$
(%i2) translate (foo)$
(%i3) foo ();
bar
throw: not within 'catch'.
 -- an error. To debug this try: debugmode(true);

Now the format of the error message in the translated case matches
the format of the error message in the interpreted case.

There was some old commented-out code in trans1 that was removed in
commit f158ab43 in 2007; it seems to suggest that at one point the
output in the interpreted case was similar to the output above.  I'm
not sure when that was: that commented-out code used an old ERLIST
macro which was obsolete even back in the Macsyma days.