Bug fix: PATCH-UP-MEVAL-IN-FSET had a bogus BARFO call
PATCH-UP-MEVAL-IN-FSET contains a call to BARFO with no arguments.
This was added in commit 122e398a in 2004, and at the time calling
BARFO with no arguments was fine.

Since commit 697a9995 in 2011, however, BARFO has required exactly
one argument.  Its lambda list only contains a rest parameter, but
passing zero arguments or more than one argument causes an error
because MFORMAT gets called with the wrong number of arguments.

Now instead of printing a message and then calling BARFO, just call
BARFO with that message as its argument.  This is what most other
calls to BARFO do.