Remove the unused FASL-DIR property from the plist on the symbol DCL
DCL used to be a transl module dealing with autoload declarations.
The FASL-DIR property was used to specify the directory that held
fasl files (on ITS only).

The ITS code that used the FASL-DIR property was removed in commit
a3d77bbc in 2007.  Most of the transl module code was removed in
commit 11be1c1f in 2011.

No functional changes.