Fix special declaration of loadprint in share/contrib/cgrind.lisp
As noted in the comment in cgrind.lisp, there was some copying from
fortra.lisp.  Based on the comment and the code it's pretty obvious
that this declaration started out as the one in fortra, but while
removing some variables from the declaration the "special" was
accidentally removed as well.

No warning was issued about this because DECLARE-TOP just ignores any
declaration identifiers besides SPECIAL and UNSPECIAL.