Remove MAP-FIND and use SOME instead
MAP-FIND was introduced in commit 19c10aef in 2015 and it is only used
in one place (another use was introduced in commit c813e416, but that
was reverted 9 days later in commit bd0f0db3).

MAP-FIND is like a single-sequence implementation of SOME, so let's
just remove it and use SOME instead.

No functional changes.

No problems with the test suite or share test suite.