Tried to compile a ChangeLog for the current maxima version.
If anyone wants to improve it feel free to do so. But if no-one starts
this tedious task we'll never have a new release. Which would be a
shame as many things have drastically improved in the meantime and also
many non-user-visible changes have taken place that make further
improvement easier by avoiding generating cryptic (because wrong)
de-facto rules.