Bug fix: compgrind:true caused compfile to not write to the output file
compgrind:true is supposed to enable pretty-printing of the translated
code that gets printed to the output file.

Originally, under Maclisp, PRINT* bound the variables ^R, ^W and
OUTFILES (and others) before calling SUB-PRINT*, which in turn would
call SPRIN1 to do the pretty-printing.  ^R and ^W were bound to T and
OUTFILES was bound to a singleton list holding the output file object;
these bindings told SPRIN1 to write to the output file and not the
terminal.

In commit e3407d64 in 2007, the call to the then-undefined SPRIN1 was
changed to just PRIN1 (without specifying any output stream).  This has
caused compfile to write the translated code to standard output instead
of the output file.