Bug fix: compgrind:true didn't ensure pretty-printing of compfile output
compgrind:true is supposed to enable pretty-printing of the translated
code that gets printed to the output file.

In commit e3407d64 in 2007, a call to the old Maclisp pretty-printer
SPRIN1 was changed to just PRIN1.  The initial value of *PRINT-PRETTY*
is implementation-dependent, so the pretty-printing of the translated
code depended on whatever value *PRINT-PRETTY* happened to have during
printing.

Now while printing the translated code we bind *PRINT-PRETTY* to true
when compgrind:true (we do this instead of calling PPRINT to avoid the
extra newline in the output, which looks strange when tr_semicompile
is true).  We specifically do not bind *PRINT-PRETTY* to false in the
case of compgrind:false (we just use whatever value it has).

See also commit a6404ac1.