Windows installer: Use current Gnuplot version (for 64 bit installers).
The most recent gnuplot installer is 5.2.8 for 64 bit (they did
not release it for 32 bit).
There was a discussion on the mailing list that keeping the same
version for 64 and 32 bit is not required and I should use the
most recent version and the versions can differ.

This fixes support request #71.