Windows installer: Use more recent VTK version (for 64 bit installers).
The newest VTK installer is 7.1.1 for 64 bit (they did not release
it for 32 bit).
There was a discussion on the mailing list that keeping the same
version for 64 and 32 bit is not required and I should use the
most recent version and the versions can differ.

This fixes ticket https://sourceforge.net/p/maxima/support-requests/70/