Fix bogus translations of atan and atan2 with float arguments
The translation of atan2 with float arguments has been inconsistent
with the interpreted case and non-float case, and the translation of
atan with a float argument has been totally broken.

The issue is that we've had a custom ATAN function that requires two
arguments and whose range is between 0 and 2*%pi instead of -%pi and
%pi.  The translations of atan and atan2 with float arguments involve
a call to this ATAN: the requirement of two arguments made our atan
translation always cause a lisp error and the different range is what
made our atan2 translation inconsistent with the other cases.

To fix this, I'm removing our custom ATAN function and we'll just use
CL:ATAN instead.  CL:ATAN takes an optional second argument and it has
a range that is consistent with our usual atan and atan2 (in fact,
CL:ATAN has been called directly in several places in src).

No problems with the test suite or share test suite.

This was discussed on the mailing list circa 2020-02-06 in the thread
"Bogus atan, atan2 translations; removing our ATAN, ATAN2".