Set the TEX-LBP property for mod
The TEX-RBP property for mod was being set twice while the TEX-LBP
property was not set at all.  mod's TEX property value is TEX-INFIX,
and setting both of the TEX-LBP and TEX-RBP properties to 180 is
consistent with what happens when calling texput with an operator
type of infix.

Discussed with Barton Willis.