translator: avoid COMPLEXIFY call if tr_float_can_branch_complex:false
When this flag is false then we assume the functions asin, acos,
etc., will not return complex numbers when passed float arguments
(and we use a mode of float when translating such calls).  Even
when we're assuming that no complex numbers will be returned, we
have been generating code to call COMPLEXIFY on the return value
of these functions.  We can generate better code that doesn't
call COMPLEXIFY in this case.

Calling COMPLEXIFY in this case doesn't make the generated code
any safer in general because if COMPLEXIFY is given a lisp complex
number (which is not a float) then it returns a Maxima complex
number (which is also not a float) -- and this violates the given
mode of float.  Some code depending on the return value from these
functions may be translated specially based on this mode, and so
returning a Maxima complex number here easily leads to errors.

tests/rtest_translator.mac tests these translations and it runs as
expected (only expected failures) with these changes.