Fix the inconsistent translation of signum with a float argument
In the interpreted case, signum returns an integer or a float when
given an integer or float argument, respectively.  In the translated
case, signum has always been returning an integer when given either
an integer or a float argument (which actually used to be consistent
with the old interpreted case in Macsyma).

Also if we translated the function

  foo (x) := signum (x)

then foo(2.0) would yield either 1 or 1.0 depending on whether or not
x was declared with mode float.

(Unrelated to the inconsistencies, the translation of signum with an
integer or float argument has been pretty silly since it reproduced
the logic of signum with every single call.)

Now we translate to a call to CL:SIGNUM when given a integer or float
argument.  This gives the desired behavior.

tests/rtest_translator.mac runs as expected, with a new test.