Fix bug #2174: Bogus translation of declare
The issue here was that declare(...) was being translated into a call
like ($DECLARE ...), which is bogus because this is a Maxima special
form.

It turns out that this actually worked in a very roundabout way under
(at least) Maclisp.  In Maclisp, if the variable UNDF-FNCTN was non-NIL
then it specified a handler function to be called in the event of a
call to an undefined function.  If the handler knew how to correct the
error, then it would return a list containing the function to be used
in place of the one in the original call (and if it could not correct
the error then the handler would just return NIL).

Macsyma installed a handler which, in the case of an MFEXPR*S like
declare, would construct and return a FEXPR which, when called, would
finally just punt to MEVAL.  For more on this, see the old UUF-HANDLER
function that was removed in commit 6a025756 in 2004.

Now we trivially translate a declare call into an appropriate call to
MEVAL (just like the translations of various other Maxima special
forms).  Having this trivial translation prevents warnings from being
printed about declare not having a full translation.

tests/rtest_translator.mac runs as expected, with a new test.