Set gamma_expand to true for some tests.
Since gamma_incomplete_lower now honors the value gamma_expand, we
need to set it to true for a couple of tests (65, 66) to expand the
expressions in the desired way to get the final simplified result.
Otherwise the test fails becaue the simplified result contains
gamma_incomplete_lower that could be expressed in simpler terms.