Use defmfun to define functions
Using defmfun instead of defun gives nice error checking for the
number of args with nice error messages.  Use it.

But note that if the error message prints out "errexp1", eval errexp1
to see the actual arg list which was too long to print out.  Or make
error_size larger.