Include keyword args when printing out function
Error messages will include the names of the keyword args.  If there's
a default value, the default value is printed.  Otherwise, a generic
"val<n>" is printed out.  Thus, you get something like

foo(a, b, [c = 99, d = val2])

Can't think of a better way to print out d=val2.  "d=" is confusing.
Maybe a short gensym?  Leave it as is for now.