No more enable GCL's GCL_MULTIPROCESS_MEMORY_POOL by default
Setting that environment variable to 't' was for a time necessary
in order to allow more than one maxima process to run on the same
machine at the same time. Setting this variable was also aggreed
to be a good idea by GCL's maintainer.

In the meantime according to the maxima-discuss mailing list on
20200411 the prefered way of enabling the multiprocess memory pool
setting has moved to the user's home directory instead and it is
suspected that setting the environment variable crashes a
gcl-compiled maxima on some computers, even if it doesn't do so on
mine.