Manual: A cross-ref from mnewton to find_root and similar functions
Often in the maxima-discuss list questions arise that seem to
indicate that the one who asks knows only mnewton or only
find_root. Let's see if that changes after this commit.
My experience until now is that questions answered in the manual
in a googleable way are only very seldomly asked in the list.