Define %%pretty-fname for use by defined functions
When printing out error messages about the wrong number of arguments,
a pretty name for the function with arguments is used.  However, this
wasn't available for use by the user.  Now, define %%PRETTY-FNAME as a
variable for use by the body of the function for printing out the
function name in the same way as printed in the argument count
checking messages.

Use this in fmin_coblya to print out nice error messages when there is
an error in the arguments to fmin_cobyla.