translator: fix bogus error reporting that caused lisp errors
*TRANSLATION-MSGS-FILES* is a list of streams to write to, but in
some MFORMAT and MGRIND calls it was incorrectly used as if it were
a stream itself.  Lisp errors would occur in these places when this
list was non-empty.

At least one stream is put on the list during translate_file and
compile_file because they want to write to the unlisp file (standard
output is also put on the list when tr_file_tty_messagesp is true).

TR-FORMAT knows how to correctly handle *TRANSLATION-MSGS-FILES*, so
change these problematic MFORMAT and MGRIND calls to TR-FORMAT calls
instead.