Clean up the implementation of ttyoff and make it reset-able
ttyoff was an alias for ^W but only "indirectly" through the variable
TTYOFF.  TTYOFF was created to be a lisp-independent way to express
whatever lisp-specific variable controlled whether or not to output
to a terminal, and on Maclisp this variable was ^W.  During the port
to Common Lisp, the Maclisp ^W just became another defined variable
in Maxima.

A very similar situation was the old writefile_on variable that was
an alias for ^R (on Maclisp) but indirectly through WRITEFILEP.  This
was cleaned up (well, removed) in commit 6bcf3cd9 in 2013.

Now define ttyoff with DEFMVAR and use it directly, and remove both
TTYOFF and ^W.

The only functional change is that ttyoff is now reset-able.

No problems with the test suite or share test suite.