Remove the translator variable transcompile
This flag controlled the output of some declarations in translated
code.  It is a relic from the days when interpreted and compiled lisp
code had different semantics.  Setting it to false would cause the
translator to generate bogus code because special declarations
required for correct semantics under Common Lisp would be omitted.

Besides the special declarations, this flag also controlled the output
of some type declarations.  I don't see a point in keeping this flag
around just to control that, so let's just remove it.

This flag was set to true in some share files, so let's remove those
uses as well.

This was discussed on the mailing list circa 2020-04-28 in the thread
"Removing transcompile".

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected.