Fix expectations for share tests for cmucl.
When running  the share testsuite for cmucl, we get the following
message:

Error summary:
Error(s) found:
  /home/toy/src/sourceforge/maxima-code/share/tensor/rtest_ctensor.mac problem:
    (34)
Tests that were expected to fail but passed:
  /home/toy/src/sourceforge/maxima-code/share/to_poly_solve/rtest_to_poly.mac problems:
    (13 14 15 16 17 18 19 20 25)
  /home/toy/src/sourceforge/maxima-code/share/to_poly_solve/rtest_to_poly_solve.mac problems:
    (199 200 201 202 203 204 214 222 268 277 278 295 311)
  /home/toy/src/sourceforge/maxima-code/share/contrib/diffequations/tests/rtest_sym.mac problems:
    (12 58)
  /home/toy/src/sourceforge/maxima-code/share/vector/rtest_vect.mac problem:
    (19)
1 test failed out of 4,360 total tests.

So update all of the expectations where cmucl passes a test that was
expected to fail.  After this fix, cmucl only fails the ctensor test,
but that's a recent failure that is being discussed on the mailing
list.