Fix bogus translations of = and # evaluations due to float contagion
The translations for evaluating = and # expressions to boolean values
(via is, if, etc.) with one float argument and a different numerical
argument (like fixnum or rational) gave wrong results because the
translator would incorrectly apply float contagion to the arguments.

This means things like is(1=1.0) and is(1/2=0.5) would yield true.

Now if we have two arguments that are both lisp numbers (modes fixnum,
float or number) then we compare them directly using EQL without doing
any type conversions.  For all other arguments we now compare them
directly using LIKE without doing any type conversions.

This is an old bug in the translator that was present in Macsyma.  This
used to translate into a = comparison instead of the EQL comparison it
has been since the port to Common Lisp (EQL didn't exist in Maclisp).
In Maclisp = gave an error if its arguments weren't both fixnums or
both flonums, so presumably this was one reason why the possible float
contagion was applied.  However, it seems like that the Maclisp EQUAL
would have been the correct predicate to use for all cases involving
lisp numbers.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected, with new tests.