Fix assignment check for ibase and obase
Both ibase and obase (*READ-BASE* and *PRINT-BASE*) can only be
integers between 2 and 36, but the check Maxima has been doing
before the assignment has allowed the symbol ROMAN.  In general
attempting to set ibase or obase to this symbol would (of course)
cause problems like a lisp error on assignment or later use.

In Maclisp IBASE and/or BASE could be set to other values, like
ROMAN for input/output in roman numerals or (only in the case of
IBASE) 1.  Macsyma under Maclisp allowed setting ibase and obase
to ROMAN but did not allow any other values that were not integers
greater than or equal to 2.

There has also been a COND clause that has done nothing since the
port to Common Lisp.  Originally if IBASE was being set then the
Maclisp sstatus option + was set either to NIL or T depending on if
the value was less than 11 or not.  This sstatus option affected how
the reader would interpret numbers.  In commit 66a62cf7 from 2007
the (conditionalized) sstatus stuff was removed.  (This clause has
affected the return value of MSETCHK but it did not affect the
assignment).

No problems with the test suite or share test suite.