Fix the bogus translation of random with a float argument
Macsyma's random was defined as a Maclisp lexpr which just called
the Maclisp RANDOM.  RANDOM took either 1 or 0 arguments and always
returned a fixnum.  In Macsyma a random call was simply translated
into a RANDOM call and this was given the mode fixnum.

random can now return either an integer or a float depending on its
(now mandatory) argument, but this mode of fixnum has remained on
the translation.  Since other code may be generated specially based
on this mode, we could get some bogus translations.  For example,
translating the following and calling foo would cause a lisp error:

  foo () := random (1.0) / 2

Now if we know that the mode of the argument is fixnum or float then
we can easily use the correct, specific mode for result.  If the mode
of the argument is not fixnum or float, then we just use a mode of
number since random always returns some type of number if it doesn't
signal an error.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected, with new tests.